Task controller updated implement policy authorization, changed the markAsDone method to status

+++ app/Http/Controllers/TaskController.php
@@ -6,15 +6,20 @@ use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
 use App\Models\Task;
 use App\Services\TaskService;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 
 class TaskController extends Controller
 {
+
+    use AuthorizesRequests;
+
+
     public function __construct(private TaskService $taskService) {}
 
-    public function index()
+    public function index(Request $request)
     {
-        $tasks = $this->taskService->list();
+        $tasks = $this->taskService->list($request);
 
         return view('tasks.index', compact('tasks'));
     }
@@ -34,26 +39,31 @@ class TaskController extends Controller
 
     public function edit(Task $task)
     {
+        $this->authorize('update', $task);
+
         return view('tasks.edit', compact('task'));
     }
 
     public function update(Task $task, UpdateTaskRequest $request)
     {
+        $this->authorize('update', $task);
+        
         $this->taskService->update($task, $request->validated());
         return redirect()->route('tasks.index')->with('success', 'Task updated.');
     }
 
     public function destroy(Task $task)
     {
+        $this->authorize('delete', $task);
         $this->taskService->delete($task);
 
         return redirect()->route('tasks.index')->with('success', 'Task deleted.');
     }
 
-    public function markDone(Task $task)
+    public function status(Task $task)
     {
-        $this->taskService->markAsDone($task);
+        $this->taskService->status($task);
 
-        return redirect()->back()->with('success, '. $task->title .  " marked as done.");
+        return redirect()->back()->with('success, ' . $task->title .  " marked as done.");
     }
 }
